[Improvement] - rework on exercise weapons system (#2703)
Rework on exercise weapon.
Added tibia global equal behavior

Added config key "maxAllowedOnADummy " default is "1"
Created table for exercise functions and variables

Start training with someone already training on that dummy and the message will be sent that someone is already training.
If someone is training and connects to the game, that user appears as Exercise Dummy Training .
If someone is training, send request to add this player on vip list, that user appears as Exercise Dummy Training.
When are training and disconnect from the game, on reconnect, the training stopped.cise Dummy Training
When are training and disconnect from the game, on reconnect, the training stopped